requirements.txt
imbalanced_learn==0.14.0
joblib==1.5.1
numpy==2.4.0
pandas==2.3.3
PyYAML==6.0.2
scikit_learn==1.8.0
matplotlib==1.3.0
